Improved the tests and changed tree structure
