+++ src/__init__.py
@@ -1,0 +1,4 @@
+# For relative imports to work in Python 3.6
+import os
+import sys
+sys.path.append(os.path.dirname(os.path.realpath(__file__)))
