+++ src/obsolete_stableDiffusionAPI.py
